Refactor big growing class file into proper modules, and did some cleanup of the required modules.
