+++ landscape_tools/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+#
+# Copyright this project and it's contributors
+# SPDX-License-Identifier: Apache-2.0
+#
+# encoding=utf8
+
+
