+++ requirements.txt
@@ -1,8 +1,5 @@
 ruamel.yaml>=0.16.10
-PyYAML>=5.3.1
 requests>=2.23.0
-validators>=0.15.0
-beautifulsoup4>=4.8.0
 url-normalize>=1.4.2
 tld>=0.12.2
 pycrunchbase>=0.3.8
